pyproject.toml
[project]
name = "mcp-agent-dev-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.123.10",
    "google-adk>=1.22.1",
    "litellm>=1.80.15",
    "ollama>=0.6.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
